1. openPdf() Added a suffix of [R] in the end of the file name.
2. closePdf()
Added a suffix of [R] in the end of the file name.

3. convertClass()
Fixed a bug from convert from factor to date, supplied an origin.

4. sketch.pptx()
a. Added additional parameters (figObj, prefix, figObjName, figList) so that it can be properly hyperlink at the footer.

5. print.figure()
To be used with export.list.of.figures()

6. sketch()
Modified ggsave default page size to landscape A4

7. export.list.of.figures()
Export list of figures into pdf or pptx.